feat: Update theme creation workflow to handle color themes
This commit modifies the `create-theme.yml` workflow file to handle color themes. It adds a new environment variable `THEME_IS_COLOR_THEME` that indicates whether the theme is a color theme. The value of this variable is determined based on the presence of the "JSON Color Theme" string in the `is-color-theme` field of the theme creation form. This change allows for the generation of a `chrome.css` file for color themes using the provided color values in a `colors.json` file.

Fixes #69