Removed list-style-image; reworked compact tabs audio icon display
List style image results in a black overlay over the mute button since a more recent Zen update. Tried a workaround with mask to get this to be transparent, but could not get it to work as intended.

Reworked the way the audio icon is displayed when tabs are compact so that it now properly hides the favicon. This also fixed an issue occurring when audio was playing from a pinned tab.